refactor: inline one-caller indirection and merge single-export modules
Removes wrappers that only delegate, files holding one alias, and layers with
a single caller. Kept every abstraction whose name still earns its keep.

trpc:
- inline checkUserAuthorized into resolveUserFromRequest, its only caller; the
  docstring claimed REST routes shared it, but nothing outside the file did
- absorb utils/headers.ts (getAuthBearer) into authorization.ts, its only
  internal consumer; it stays exported for apps/web
- merge procedures.ts and middlewares/verifyAuthentication.ts into trpc.ts, so
  the context, the t instance, and protectedProcedure live in one file
- drop ITRPCContext and the explicit return annotation; the context type is
  already derived via Awaited<ReturnType<typeof createTRPCContext>>
- drop the one-member EFirebaseDBRootKeys enum and the union no caller could
  satisfy; inline mapExtension, which made both GitHub type aliases dead

shared:
- move the three z.infer aliases next to their schemas and delete the
  interfaces/ directory; the barrel re-exports the types explicitly
- fold Bookmarks/mapper into Bookmarks/utils (same barrel export)
- merge revokeBlobUrl into evictBlobUrl, and drop getBlobUrlFromCache's
  duplicate of the cache-hit check getBlobUrlFromOpenCache already does
- drop asyncNoOp, used once

apps:
- give each app its own one-line getFaviconUrl instead of aliasing a shared
  helper that existed only to be renamed
- inline verifyInternalToken into the storage-cleanup route, which now answers
  403 rather than throwing a bare Error that surfaced as a 500. The cron
  workflow greps the success body, which is unchanged.
- move openNewTab into its only consumer, drop the Set-of-one RESTRICTED_PATHS
  and the one-interface feature.ts
- merge the web base-fixture into auth-fixture and inline the DownloadPage
  page object, whose one method held two assertions

Also restores TEST_SHORTCUTS.MANTINE/TWITCH/YOUTUBE, wrongly dropped in
d4f81097: shortcuts.spec asserts the panel's rule count against
Object.keys(TEST_SHORTCUTS).length, so entries with no member reference are
still load bearing. Comment added so they are not re-dropped.

Verified: lint, format, typecheck, build, 93 extension and 24 web e2e tests.
