+++ apps/extension/src/constants/favicon.ts
@@ -1,3 +1,2 @@
-import { getGoogleFaviconUrl } from '@bypass/shared';
-
-export const getFaviconUrl = getGoogleFaviconUrl;
+export const getFaviconUrl = (url: string) =>
+  `https://www.google.com/s2/favicons?domain=${new URL(url).hostname}&sz=32`;
