--- apps/web/src/app/components/types/feature.ts
@@ -1,7 +1,0 @@
-import type { IconSvgElement } from '@hugeicons/react';
-
-export default interface Feature {
-  icon: IconSvgElement;
-  title: string;
-  content: string;
-}
