--- apps/web/src/app/utils/index.ts
@@ -1,4 +1,0 @@
-export const openNewTab = (url: string) => {
-  // `noopener` makes window.open return null by spec, so there is no handle to focus
-  window.open(url, '_blank', 'noopener,noreferrer');
-};
