--- apps/web/tests/page-object-models/download-page.ts
@@ -1,12 +1,0 @@
-import { expect, type Page } from '@playwright/test';
-
-export class DownloadPage {
-  constructor(readonly page: Page) {}
-
-  async testPageMetaData() {
-    await expect(this.page).toHaveTitle('Bypass Links');
-    await expect(this.page.getByRole('heading', { level: 1 })).toHaveText(
-      'Skip the Wait. Bypass Links Instantly.'
-    );
-  }
-}
