--- packages/shared/src/components/Bookmarks/mapper/index.ts
@@ -1,31 +1,0 @@
-import {
-  type ContextBookmark,
-  type IBookmarksObj,
-  type IFolderMetaData,
-} from '../interfaces';
-import { getDecryptedBookmark, getDecryptedFolder } from '../utils';
-
-export const bookmarksMapper = (
-  { isDir, hash }: IFolderMetaData,
-  urlList: IBookmarksObj['urlList'],
-  folderList: IBookmarksObj['folderList']
-): ContextBookmark => {
-  if (isDir) {
-    const folder = getDecryptedFolder(folderList[hash]);
-    return {
-      id: folder.id,
-      isDir,
-      name: folder.name,
-      isDefault: Boolean(folder.isDefault),
-    };
-  }
-
-  const bookmark = getDecryptedBookmark(urlList[hash]);
-  return {
-    id: bookmark.id,
-    isDir,
-    url: bookmark.url,
-    title: bookmark.title,
-    taggedPersons: bookmark.taggedPersons,
-  };
-};
