--- packages/shared/src/interfaces/ILastVisited.ts
@@ -1,5 +1,0 @@
-import { type z } from 'zod/mini';
-
-import { type LastVisitedSchema } from '../schema/lastVisitedSchema';
-
-export type ILastVisited = z.infer<typeof LastVisitedSchema>;
