--- packages/shared/src/interfaces/IRedirection.ts
@@ -1,10 +1,0 @@
-import { type z } from 'zod/mini';
-
-import {
-  type RedirectionSchema,
-  type RedirectionsSchema,
-} from '../schema/redirectionSchema';
-
-export type IRedirection = z.infer<typeof RedirectionSchema>;
-
-export type IRedirections = z.infer<typeof RedirectionsSchema>;
