--- packages/shared/src/interfaces/IWebsites.ts
@@ -1,5 +1,0 @@
-import { type z } from 'zod/mini';
-
-import { type WebsitesSchema } from '../schema/websitesSchema';
-
-export type IWebsites = z.infer<typeof WebsitesSchema>;
