+++ packages/shared/src/schema/lastVisitedSchema.ts
@@ -1,3 +1,5 @@
 import { z } from 'zod/mini';
 
 export const LastVisitedSchema = z.record(z.string(), z.number());
+
+export type ILastVisited = z.infer<typeof LastVisitedSchema>;
