+++ packages/trpc/src/@types/trpc.ts
@@ -1,13 +1,7 @@
-import { type AuthorizationResult } from '../utils/authorization';
-
 export interface IUser {
   readonly uid: string;
   readonly email?: string;
   readonly emailVerified: boolean;
   readonly displayName?: string;
   readonly photoURL?: string;
-}
-
-export interface ITRPCContext {
-  auth: AuthorizationResult;
 }
