+++ packages/trpc/src/index.ts
@@ -1,6 +1,5 @@
 export { createTRPCContext } from './trpc';
 export * from './routers';
-export { checkUserAuthorized } from './utils/authorization';
-export { getAuthBearer } from './utils/headers';
+export { getAuthBearer } from './utils/authorization';
 
 export { cleanupStorage } from './services/storageCleanupService';
