--- packages/trpc/src/procedures.ts
@@ -1,4 +1,0 @@
-import verifyAuthMiddleware from './middlewares/verifyAuthentication';
-import { t } from './trpc';
-
-export const protectedProcedure = t.procedure.use(verifyAuthMiddleware);
