+++ packages/trpc/src/trpc.ts
@@ -1,11 +1,8 @@
-import { initTRPC } from '@trpc/server';
+import { initTRPC, TRPCError } from '@trpc/server';
 
-import { type ITRPCContext } from './@types/trpc';
 import { resolveUserFromRequest } from './utils/authorization';
 
-export const createTRPCContext = async (
-  req: Request
-): Promise<ITRPCContext> => {
+export const createTRPCContext = async (req: Request) => {
   const auth = await resolveUserFromRequest(req);
 
   return { auth };
@@ -14,3 +11,18 @@ export const createTRPCContext = async (
 export const t = initTRPC
   .context<Awaited<ReturnType<typeof createTRPCContext>>>()
   .create();
+
+export const protectedProcedure = t.procedure.use(async (opts) => {
+  const { ctx } = opts;
+
+  if (!ctx.auth.ok) {
+    throw new TRPCError({
+      code: ctx.auth.status === 401 ? 'UNAUTHORIZED' : 'FORBIDDEN',
+      message: ctx.auth.message,
+    });
+  }
+
+  return opts.next({
+    ctx: { ...ctx, user: ctx.auth.user }, // For type safety in protected procedures
+  });
+});
