+++ packages/trpc/src/utils/firebase.ts
@@ -1,10 +1,7 @@
 import filenamify from 'filenamify';
 
-import { EFirebaseDBRootKeys } from '../constants/firebase';
-
-export const getFullDbPath = (ref: string, uid?: string) => {
-  return `/${EFirebaseDBRootKeys.data}/${uid}/${ref}`;
-};
+export const getFullDbPath = (ref: string, uid?: string) =>
+  `/data/${uid}/${ref}`;
 
 export const getBucketPath = (uid: string) => `${uid}/persons`;
