--- packages/trpc/src/utils/headers.ts
@@ -1,2 +1,0 @@
-export const getAuthBearer = (req: Request) =>
-  req.headers.get('authorization')?.split?.('Bearer ')?.[1];
